Kill off LP64-related warnings.

+++ xc/programs/Xserver/record/set.c
@@ -199,7 +199,7 @@ BitVectorIterateSet(pSet, pIter, pInterval)
     RecordSetIteratePtr pIter;
     RecordSetInterval *pInterval;
 {
-    int iterbit = (int)pIter;
+    int iterbit = (int)(long)pIter;
     int b;
 
     b = BitVectorFindBit(pSet, iterbit, TRUE);
@@ -208,7 +208,7 @@ BitVectorIterateSet(pSet, pIter, pInterval)
 
     b = BitVectorFindBit(pSet, b, FALSE);
     pInterval->last = (b < 0) ? ((BitVectorSetPtr)pSet)->maxMember : b - 1;
-    return (RecordSetIteratePtr)(pInterval->last + 1);
+    return (RecordSetIteratePtr)(long)(pInterval->last + 1);
 }
 
 RecordSetOperations BitVectorSetOperations = {
@@ -517,7 +517,7 @@ RecordCreateSet(pIntervals, nIntervals, pMem, memsize)
 					&pCreateSet);
     if (pMem)
     {
-	if ( ((int)pMem & (alignment-1) ) || memsize < size)
+	if ( ((long)pMem & (alignment-1) ) || memsize < size)
 	    return NULL;
     }
     return (*pCreateSet)(pIntervals, nIntervals, pMem, size);
